projects/memagents/pyproject.toml
[project]
name = "memagents"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.13",
    "ipykernel>=6.29.5",
    "llama-index-core>=0.14.13",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-llms-openai>=0.6.15",
    "llama-index-memory-mem0>=1.0.0",
    "llama-index-vector-stores-chroma>=0.5.5",
    "llama-index-vector-stores-qdrant>=0.9.1",
    "mem0ai>=0.1.114",
    "openai>=1.93.0",
    "protobuf",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.23.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.3",
]

[tool.setuptools]
package-dir = {"" = "asdrp"}

[tool.setuptools.packages.find]
where = ["asdrp"]

[tool.uv.workspace]
members = [
]

